refactor: extract electron related files
